WinRT: Fix build for Windows 10
bf24838c3344f009f9fe40f596a4eab798f071b3 introduced a hard dependency on
kernel32.lib, which needs to be added for the msvc2015 mkspecs.

As the library differs for Windows Phone / Windows 10 Mobile, this most
likely has the side-effect that we will need to introduce winphone-xxx-
msvc2015 at a later stage.

Change-Id: I8fa2a68d421345c14ba90efc6faa5eea1ad457e1
Reviewed-by: Andrew Knight <andrew.knight@intopalo.com>