Fix surface resize
If you call QWindowSystemInterface::handleExposeEvent() from the render
thread you won't give a change for the main thread to process further
resize events. In a long resize animation (like dragging a window border
with mouse) the resize events will queue up like crazy, leving
handleSurfaceResize() to deal with a ton of outdated resize events
once the resize animation is finally over.

Furthermore, as there's no synchronicity between the processing of resize
events and the consupmtion of buffers, there's no point in trying to tie
one to the other in any way. So better ignore the actual size information
from a resize event.

Change-Id: I5109245761f19a527d98c3c6f5d24789b633ed14
Reviewed-by: Gerry Boland <gerry.boland@canonical.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>