PPM image format: avoid useless processing on truncated files
For ascii formatted files, the handler would not stop processing on a
premature EOF, but instead continue to fill the whole image memory
area byte by byte. This could lead to unexpected CPU exhaustion in the
case of a small image file declaring huge image dimensions, but having
no content.

Instead, check for EOF for each scanline and quit processing if found.

Change-Id: I8dbcd7eb34553873e49706d61b5752f22e04eb6a
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>