iOS: Make setFrame in desktop view take superview transforms into account
An assumption we do for the QIOSDesktopManagerView is that it should always
cover the whole UIWindow. To achieve that, we override setFrame to be stop
any attempt from UIKit to make our view smaller, e.g. when the statusbar
changes height.

In case the view is not a direct child of the window, we need to take
any transformations into account when computing the new frame. This
happens e.g. during presentation of other view-controllers, where our
view is temporarily reparented into a UITransitionView that may have
a transform set.

Task-number: QTBUG-47506
Change-Id: I388143f2cbb566541ffb1068443ce21e62ea2b42
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@theqtcompany.com>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@theqtcompany.com>